backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
gitpython==3.1.40
groq==0.4.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-socketio==5.10.0
aiofiles==23.2.1
requests==2.31.0
beautifulsoup4==4.12.2
langchain==0.1.0
langchain-groq==0.0.1
langgraph==0.0.20
google-search-results==2.4.2